feat: Add firstTimeLogin field to User and TransactionParty models
This commit adds the `firstTimeLogin` field to the `User` and `TransactionParty` models in the Prisma schema. The field is set to `true` by default and indicates whether it is the user's first time logging in. The field is also added to the `User` and `TransactionParty` tables in the database through a migration.

Please review and merge this commit.