chore: Update login response in AuthService
The code changes in `auth.service.ts` update the login response in the `AuthService` class. If `company.firstTimeLogin` is true, the `firstTimeLogin` field is set to false and the `company` object is updated in the database. The login response now includes the `firstTimeLogin` field set to true. If `company.firstTimeLogin` is false, the `firstTimeLogin` field is set to false in the login response.

Please review and merge this commit.