feat: Update login response and add firstTimeLogin field to User and TransactionParty models
This commit updates the login response in the `AuthService` class to include the `firstTimeLogin` field. If `company.firstTimeLogin` is true, the field is set to true in the login response. If `company.firstTimeLogin` is false, the field is set to false in the login response. Additionally, the commit adds the `firstTimeLogin` field to the `User` and `TransactionParty` models in the Prisma schema. The field is set to `true` by default and indicates whether it is the user's first time logging in. The field is also added to the `User` and `TransactionParty` tables in the database through a migration.

Please review and merge this commit.