feat: Add check for PII status in user dashboard data
This commit adds a check for the status of Personally Identifiable Information (PII) in the user dashboard data. It retrieves the PII status for the user from the `basic_pii` and `secret_pii` tables in the database. If both PII statuses are not set, it updates the dashboard data to indicate that the basic PII and secret PII are not set. Otherwise, it updates the dashboard data to indicate that the basic PII and secret PII are set.

Please review and merge this commit.