refactor: Update PII status in user dashboard data
This commit refactors the code in the `UserService` class to update the Personally Identifiable Information (PII) status in the user dashboard data. Instead of using string values like 'Basic PII Not Set' and 'Secret PII Not Set', the code now uses boolean values to indicate whether the basic PII and secret PII are saved. If both PII statuses are not set, the `basic_pii_saved` and `secret_pii_saved` fields in the dashboard data are set to `false`. Otherwise, they are set to `true`.